+++ NEWS.rst
@@ -1,5 +1,14 @@
 .. default-role:: code
 
+Unreleased
+======================================================================
+
+Supports Python 3.x – Python 3.y
+
+New Features
+------------------------------
+* `setv` now supports chained assignment with `:chain`.
+
 1.1.0 ("Business Hugs", released 2025-05-08)
 ======================================================================